Split config into 3 files for each subproject -> Rainfall models/API/Webapp | Preparing project split!
