+++ run.py
@@ -18,7 +18,7 @@ def run():
 @run.command()
 @click.pass_context
 def api(ctx):
-    from config import Config
+    from back.api.config import Config
 
     uvicorn.run(
         "back.api.app:fastapi_app",
@@ -29,8 +29,8 @@ def api(ctx):
 @run.command()
 @click.pass_context
 def webapp(ctx):
-    from config import Config
     from webapp.app import flask_app
+    from webapp.config import Config
 
     flask_app.run(**ctx.ensure_object(Config).get_webapp_server_settings.model_dump())
