+++ tst/test_config.py
@@ -1,14 +1,19 @@
 from pytest import fixture, raises
 from yaml.parser import ParserError  # type: ignore
 
-from config import APISettings, Config, WebappServerSettings
-
-config = Config()
+from back.api.config import APISettings
+from back.api.config import Config as APIConfig
+from back.rainfall.config import Config as RainfallConfig
+from base_config import BaseConfig
+from webapp.config import Config as WebappConfig
+from webapp.config import WebappServerSettings
 
 
 @fixture(autouse=True)
 def reset_config():
-    Config._instance = None
+    for config in [BaseConfig, RainfallConfig, APIConfig, WebappConfig]:  # type: ignore
+        config._instance = None
+
     yield
 
 
@@ -16,32 +21,28 @@ class TestConfig:
     @staticmethod
     def test_config_file_not_found():
         with raises(FileNotFoundError):
-            Config(path="/there/is/no/config/file/there/dude.yaml")
+            BaseConfig(path="/there/is/no/config/file/there/dude.yaml")
 
     @staticmethod
     def test_config_file_invalid():
         with raises(ParserError):
-            Config(path="README.md")
+            BaseConfig(path="README.md")
 
     @staticmethod
-    def test_get_dataset_url():
-        assert isinstance(config.get_data_settings.file_url, str)
+    def test_get_data_settings():
+        data_settings = RainfallConfig().get_data_settings
 
-    @staticmethod
-    def test_get_dataset_path():
-        assert isinstance(config.get_data_settings.local_file_path, str)
+        assert isinstance(data_settings.file_url, str)
 
-    @staticmethod
-    def test_get_start_year():
-        assert isinstance(config.get_data_settings.start_year, int)
+        if data_settings.local_file_path is not None:
+            assert isinstance(data_settings.local_file_path, str)
 
-    @staticmethod
-    def test_get_rainfall_precision():
-        assert isinstance(config.get_data_settings.rainfall_precision, int)
+        assert isinstance(data_settings.start_year, int)
+        assert isinstance(data_settings.rainfall_precision, int)
 
     @staticmethod
     def test_get_api_server_settings():
-        api_server_settings = config.get_api_settings.server
+        api_server_settings = APIConfig().get_api_settings.server
 
         assert isinstance(api_server_settings, APISettings.APIServerSettings)
         assert api_server_settings.model_fields.keys() == {
@@ -52,8 +53,10 @@ class TestConfig:
 
     @staticmethod
     def test_get_webapp_server_settings():
-        assert isinstance(config.get_webapp_server_settings, WebappServerSettings)
-        assert config.get_webapp_server_settings.model_fields.keys() == {
+        webapp_server_settings = WebappConfig().get_webapp_server_settings
+
+        assert isinstance(webapp_server_settings, WebappServerSettings)
+        assert webapp_server_settings.model_fields.keys() == {
             "host",
             "port",
             "debug",
