+++ webapp/config.yml
@@ -1,0 +1,4 @@
+webapp:  # Flask
+  host: 127.0.0.1
+  port: 5000
+  debug: true
